chore: Improve MainActivity readability by moving composable functions to dedicated files
- Add MainScreen to declare MainActivity content
- Add MainBottomNavigationBar to display the bottom navigation bar for root screens (Popular, TopRated, Favorite)
- Add Navigation screen to declare navigation graphs
